requirements.txt
Flask==1.1.1
python-dotenv==0.12.0
Flask-RESTful==0.3.8